Modified version of Potentiometer_to_servo
Essentially this is Potentiometer_to_servo_v2. This is the first implementation with two buttons added as well as a buzzer that is supposed to beep based on if the right combination was entered or not.

Whats missing:
The button combination, currently you can only hit the two buttons once each then it "unlocks" but the servo doesn't stay unlocked so you need to enter the buttons again before the servo will move again.
The buzzer also needs to be worked on. Right now it doesn't work.